+++ .ci/script.sh
@@ -1,9 +1,35 @@
 #!/usr/bin/env bash
 set -ex
 
-num_threads=4
+# Sanity checks for required environment variables.
+if [ -z "$BUILD_TYPE" ]; then
+  echo "Error: Environment variable BUILD_TYPE is unset."
+  exit 1
+fi
 
-# https://unix.stackexchange.com/a/129401
+if [ -z "$BUILD_DARTPY" ]; then
+  echo "Info: Environment variable BUILD_DARTPY is unset. Using OFF by default."
+  BUILD_DARTPY=OFF
+fi
+
+if [ -z "$BUILD_DOCS" ]; then
+  echo "Info: Environment variable BUILD_DOCS is unset. Using OFF by default."
+  BUILD_DOCS=OFF
+fi
+
+if [ -z "$CODECOV" ]; then
+  echo "Info: Environment variable CODECOV is unset. Using OFF by default."
+  CODECOV=OFF
+fi
+
+if [ -z "$OS_NAME" ]; then
+  echo "Error: Environment variable OS_NAME is unset."
+  exit 1
+fi
+
+# Set number of threads for parallel build
+# Ref: https://unix.stackexchange.com/a/129401
+num_threads=4
 while getopts ":j:" opt; do
   case $opt in
     j) num_threads="$OPTARG"
@@ -13,12 +39,15 @@ while getopts ":j:" opt; do
   esac
 done
 
+# Set compilers
 if [ "$COMPILER" = "gcc" ]; then
   export CC=gcc
   export CXX=g++
 elif [ "$COMPILER" = "clang" ]; then
   export CC=clang
   export CXX=clang++
+else
+  echo "Info: Compiler isn't specified. Using the system default."
 fi
 
 # Skip Xenial and Bionic in push builds
@@ -35,10 +64,20 @@ fi
 
 mkdir build && cd build
 
-if [ "$BUILD_NAME" = "TRUSTY_DEBUG" ]; then
-  cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DDART_VERBOSE=ON -DDART_TREAT_WARNINGS_AS_ERRORS=ON -DDART_BUILD_EXTRAS=ON -DDART_CODECOV=ON ..
-else
-  cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DDART_VERBOSE=ON -DDART_TREAT_WARNINGS_AS_ERRORS=ON -DDART_BUILD_EXTRAS=ON -DDART_CODECOV=OFF ..
+cmake .. \
+  -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
+  -DDART_BUILD_DARTPY=$BUILD_DARTPY \
+  -DDART_VERBOSE=ON \
+  -DDART_TREAT_WARNINGS_AS_ERRORS=ON \
+  -DDART_BUILD_EXTRAS=ON \
+  -DDART_CODECOV=$CODECOV
+
+if [ "$BUILD_DARTPY" = "ON" ]; then
+  make -j$num_threads binding
+  make -j$num_threads dartpy
+
+  # Disabled for now
+  # make pytest
 fi
 
 if [ "$OS_NAME" = "linux" ]; then
@@ -47,7 +86,7 @@ else
   make -j$num_threads all tests
 fi
 
-if [ "$OS_NAME" = "linux" ] && [ $(lsb_release -sc) = "trusty" ]; then
+if [ "$OS_NAME" = "linux" ] && [ $(lsb_release -sc) = "bionic" ]; then
   make check-format
 fi
 
@@ -65,8 +104,3 @@ cd $BUILD_DIR/examples/rigidCubes
 mkdir build && cd build
 cmake ..
 make -j4
-
-# Uploading report to CodeCov
-if [ "$BUILD_NAME" = "TRUSTY_DEBUG" ]; then
-  bash <(curl -s https://codecov.io/bash) || echo "Codecov did not collect coverage reports"
-fi