+++ dart/constraint/BoxedLcpConstraintSolver.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2018, The DART development contributors
+ * Copyright (c) 2011-2019, The DART development contributors
  * All rights reserved.
  *
  * The list of contributors can be found at:
@@ -160,12 +160,8 @@ void BoxedLcpConstraintSolver::solveConstrainedGroup(ConstrainedGroup& group)
       }
     }
 
-    assert(
-        isSymmetric(
-            n,
-            mA.data(),
-            mOffset[i],
-            mOffset[i] + constraint->getDimension() - 1));
+    assert(isSymmetric(
+        n, mA.data(), mOffset[i], mOffset[i] + constraint->getDimension() - 1));
 
     constraint->unexcite();
   }